// eshop-ui-customer-portal/src/app/services/api/catalog.api.client.ts
export class CatalogBrandReadModel implements ICatalogBrandReadModel {
    ts?: string;
    id?: string;
    brand?: string;

    constructor(data?: ICatalogBrandReadModel) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.ts = _data["ts"];
            this.id = _data["id"];
            this.brand = _data["brand"];
        }
    }

    static fromJS(data: any): CatalogBrandReadModel {
        data = typeof data === 'object' ? data : {};
        let result = new CatalogBrandReadModel();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["ts"] = this.ts;
        data["id"] = this.id;
        data["brand"] = this.brand;
        return data;
    }
}

export interface ICatalogBrandReadModel {
    ts?: string;
    id?: string;
    brand?: string;
}
// ...
export class CatalogItemReadModel implements ICatalogItemReadModel {
    ts?: string;
    id?: string;
    name?: string;
    description?: string | undefined;
    price?: number | undefined;
    pictureUri?: string | undefined;
    catalogTypeId?: string;
    catalogBrandId?: string;
    catalogType?: CatalogTypeReadModel;
    catalogBrand?: CatalogBrandReadModel;
    availableQty?: number;

    constructor(data?: ICatalogItemReadModel) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.ts = _data["ts"];
            this.id = _data["id"];
            this.name = _data["name"];
            this.description = _data["description"];
            this.price = _data["price"];
            this.pictureUri = _data["pictureUri"];
            this.catalogTypeId = _data["catalogTypeId"];
            this.catalogBrandId = _data["catalogBrandId"];
            this.catalogType = _data["catalogType"] ? CatalogTypeReadModel.fromJS(_data["catalogType"]) : <any>undefined;
            this.catalogBrand = _data["catalogBrand"] ? CatalogBrandReadModel.fromJS(_data["catalogBrand"]) : <any>undefined;
            this.availableQty = _data["availableQty"];
        }
    }

    static fromJS(data: any): CatalogItemReadModel {
        data = typeof data === 'object' ? data : {};
        let result = new CatalogItemReadModel();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["ts"] = this.ts;
        data["id"] = this.id;
        data["name"] = this.name;
        data["description"] = this.description;
        data["price"] = this.price;
        data["pictureUri"] = this.pictureUri;
        data["catalogTypeId"] = this.catalogTypeId;
        data["catalogBrandId"] = this.catalogBrandId;
        data["catalogType"] = this.catalogType ? this.catalogType.toJSON() : <any>undefined;
        data["catalogBrand"] = this.catalogBrand ? this.catalogBrand.toJSON() : <any>undefined;
        data["availableQty"] = this.availableQty;
        return data;
    }
}

export interface ICatalogItemReadModel {
    ts?: string;
    id?: string;
    name?: string;
    description?: string | undefined;
    price?: number | undefined;
    pictureUri?: string | undefined;
    catalogTypeId?: string;
    catalogBrandId?: string;
    catalogType?: CatalogTypeReadModel;
    catalogBrand?: CatalogBrandReadModel;
    availableQty?: number;
}

export class CatalogTypeReadModel implements ICatalogTypeReadModel {
    ts?: string;
    id?: string;
    type?: string;

    constructor(data?: ICatalogTypeReadModel) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.ts = _data["ts"];
            this.id = _data["id"];
            this.type = _data["type"];
        }
    }

    static fromJS(data: any): CatalogTypeReadModel {
        data = typeof data === 'object' ? data : {};
        let result = new CatalogTypeReadModel();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["ts"] = this.ts;
        data["id"] = this.id;
        data["type"] = this.type;
        return data;
    }
}

export interface ICatalogTypeReadModel {
    ts?: string;
    id?: string;
    type?: string;
}